Reorganization, code toggle for every page
