+++ gatsby-ssr.js
@@ -7,8 +7,8 @@
 import React, { Component } from "react";
 import { CssBaseline, ThemeProvider } from "@material-ui/core";
 // import { StylesProvider } from "@material-ui/styles";
-import Topbar from "./src/components/Topbar";
-import Footer from "./src/components/Footer";
+import Topbar from "./src/components/Layout/components/Topbar";
+import Footer from "./src/components/Layout/components/Footer";
 import theme from "./src/styles/theme";
 import "./src/styles/particlesStyles.css";
