+++ src/components/About/components/AboutCard.js
@@ -9,7 +9,7 @@ import {
   Divider,
 } from "@material-ui/core";
 import PropTypes from "prop-types";
-import aboutCardStyles from "../../styles/aboutCardStyles";
+import aboutCardStyles from "../../../styles/aboutCardStyles";
 
 //todo: weird corner clipping w/solid main background
