+++ src/components/Layout/components/Topbar.js
@@ -16,7 +16,7 @@ import BuildIcon from "@material-ui/icons/Build";
 import FindInPageIcon from "@material-ui/icons/FindInPage";
 import CodeIcon from "@material-ui/icons/Code";
 import CropOriginalIcon from "@material-ui/icons/CropOriginal";
-import topbarStyles from "../styles/topbarStyles";
+import topbarStyles from "../../../styles/topbarStyles";
 
 //todo: maybe indicator for which page youre on, just extract from path
 //todo: maybe change when mobile buttons appear. when vertical view starts @750. probably use fade since hidden doesnt have a true/false
