+++ src/pages/projects.js
@@ -1,112 +1,43 @@
 import React, { Component } from "react";
-import {
-  Fade,
-  withStyles,
-  Grid,
-  Typography,
-  //   Card,
-  //   CardContent,
-  //   Divider,
-  //   CardActions,
-} from "@material-ui/core";
-import Layout from "../components/layout";
-import PropTypes from "prop-types";
-import KeyboardArrowDownIcon from "@material-ui/icons/KeyboardArrowDown";
-import projectsStyles from "../styles/projectsStyles";
+import Layout from "../components/Layout/Layout";
+import ProjectsView from "../components/Projects/ProjectsView";
+import TabbedCodeView from "../components/TabbedCodeView";
 
-const cron = require("node-cron");
-
-//todo: give placeholder height for loading element so the page doesnt collapse randomly on refresh
+//todo: move away from br as spacing, just calculate how many pixels its taking up instead and use that for margin.
+//todo: get rid of unecessary react.fragment
+//todo: solution is to pass a callback into layout and have layout call it, but i dont want to make a function in every page for it.
+//important: will not rerender if props are same.
+//todo: maybe move layout back into the wrapper, research more about wrapRoot vs wrapPage. if it doesnt get rerendered at all, keep it in layout. also maybe move particles into layout?
 
 class Projects extends Component {
-  constructor(props) {
-    super(props);
+  state = { codeView: false };
 
-    this.state = {
-      showIntro: false,
-      showArrow: false,
-    };
+  toggleCodeView = (codeView) => {
+    this.setState({ codeView });
+  };
 
-    this.flashArrow = cron.schedule(
-      "*/1 * * * * *",
-      () => {
-        this.setState({ showArrow: !this.state.showArrow });
-      },
-      { scheduled: false }
-    );
-  }
+  renderView = () => {
+    let codeView = false;
 
-  componentDidMount = () => {
-    setTimeout(() => {
-      this.setState({ showIntro: true });
-    }, 1000);
+    if (typeof localStorage !== "undefined") {
+      codeView = JSON.parse(localStorage.getItem("codeView")) || false;
+    }
 
-    setTimeout(() => {
-      this.flashArrow.start();
-    }, 2000);
+    if (!codeView) {
+      console.log("Rendering bio");
+      return <ProjectsView />;
+    } else {
+      console.log("Rendering code view");
+      return <TabbedCodeView />;
+    }
   };
 
   render() {
-    const classes = this.props.classes;
-
     return (
-      <Layout>
+      <Layout toggleCodeView={this.toggleCodeView}>
         <br />
         <br />
-        <Grid
-          container
-          direction="column"
-          justify="center"
-          alignItems="center"
-          className={classes.root}
-        >
-          <Grid item>
-            <Grid
-              container
-              direction="row"
-              justify="center"
-              alignItems="center"
-            >
-              <Grid item>
-                <Fade in={this.state.showIntro} timeout={3000}>
-                  <Typography variant="h3" className={classes.greeting}>
-                    Explore some of my projects
-                  </Typography>
-                </Fade>
-              </Grid>
-            </Grid>
-          </Grid>
-          <Grid item>
-            <Grid
-              container
-              spacing={0}
-              direction="row"
-              justify="center"
-              alignItems="center"
-            >
-              <Grid item>
-                <Fade in={this.state.showArrow} timeout={990}>
-                  <KeyboardArrowDownIcon fontSize="large" />
-                </Fade>
-              </Grid>
-            </Grid>
-          </Grid>
-          <br />
-          <br />
-          <Grid item>
-            <Grid
-              container
-              spacing={0}
-              direction="row"
-              justify="center"
-              alignItems="center"
-            >
-              <Grid item>
-                <h1>projects go here</h1>
-              </Grid>
-            </Grid>
-          </Grid>
-        </Grid>
+        {this.renderView()}
         <br />
         <br />
         <br />
@@ -116,8 +47,4 @@ class Projects extends Component {
   }
 }
 
-Projects.propTypes = {
-  className: PropTypes.string,
-};
-
-export default withStyles(projectsStyles)(Projects);
+export default Projects;
