+++ src/styles/aboutViewStyles.js
