--- src/styles/appStyles.js
@@ -1,15 +1,0 @@
-export default (theme) => ({
-  particleBG: {
-    width: "100%",
-    height: "100%",
-    position: "fixed",
-    margin: 0,
-    padding: 0,
-    top: 0,
-    left: 0,
-    zIndex: -999,
-    backgroundColor: "rgb(51, 174, 255)",
-  },
-});
-
-//todo: doesnt work. fix it man. unless you wanna keep the css
