+++ src/styles/projectsViewStyles.js
